vt_web/requirements.txt
django~=5.2
django-configurations
gunicorn
pillow
django-allauth[socialaccount,openid_connect]
requests
jwt
django-bootstrap5

djangorestframework
djangorestframework-api-key~=3.1.0
django-mathfilters

django-simple-captcha
python-decouple